feat: refactor scouting page to use query parameters for team selection and improve component structure
